Extract auth cookies and zitadel config to submodules
Convert backend/routes/auth.py to a package. Two module-level surfaces move
out of the giant register_auth_routes file:

- backend/routes/auth/cookies.py: signed-cookie serializers and the five
  set/load/clear helper trios for the auth flows.
- backend/routes/auth/zitadel_config.py: env-driven ZITADEL/OAuth config
  getters, id-token decoder, and PKCE challenge builder.

__init__.py drops from 3327 to 2965 LOC. Pure relocation; behavior is
unchanged.
