+++ backend/tests/test_runtime_typing_guards.py
@@ -6,7 +6,9 @@ _RUNTIME_FILES = [
     "backend/app.py",
     "backend/models/auth.py",
     "backend/models/main_db.py",
-    "backend/routes/auth.py",
+    "backend/routes/auth/__init__.py",
+    "backend/routes/auth/cookies.py",
+    "backend/routes/auth/zitadel_config.py",
     "backend/routes/agreements.py",
     "backend/routes/reference_data.py",
     "backend/routes/sections.py",
